+++ Utils/src/main/java/com/distrimind/util/crypto/P2PJPAKESecretMessageExchanger.java
@@ -78,13 +78,13 @@ public class P2PJPAKESecretMessageExchanger extends Agreement {
 	private char[] getHashedPassword(char[] message, byte salt[], int offset_salt, int len_salt) throws NoSuchAlgorithmException, InvalidKeySpecException, NoSuchProviderException
 	{
 		byte[] m = hashMessage(MessageDigestType.BC_FIPS_SHA3_256.getMessageDigestInstance(), message, salt,
-				offset_salt, len_salt, PasswordHashType.BCRYPT, 15);
+				offset_salt, len_salt, PasswordHashType.BCRYPT, (byte)15);
 		return convertToChar(m);
 	}
 	private char[] getHashedPassword(byte[] message, int offset, int len, byte salt[], int offset_salt, int len_salt, boolean messageIsKey) throws NoSuchAlgorithmException, InvalidKeySpecException, NoSuchProviderException
 	{
 		byte[] m = hashMessage(MessageDigestType.BC_FIPS_SHA3_256.getMessageDigestInstance(), message, offset, len,
-				salt, offset_salt, len_salt, messageIsKey ? null : PasswordHashType.BC_PBKFD2WithHMacSHA3_512, messageIsKey?10:15);
+				salt, offset_salt, len_salt, messageIsKey ? null : PasswordHashType.BC_PBKFD2WithHMacSHA3_512, messageIsKey?(byte)10:(byte)15);
 		return convertToChar(m);
 	}
 	
@@ -136,7 +136,7 @@ public class P2PJPAKESecretMessageExchanger extends Agreement {
 	}
 
 	private static byte[] hashMessage(AbstractMessageDigest messageDigest, byte data[], int off, int len, byte[] salt,
-			int offset_salt, int len_salt, PasswordHashType passwordHashType, int cost)
+			int offset_salt, int len_salt, PasswordHashType passwordHashType, byte cost)
 			throws NoSuchAlgorithmException, InvalidKeySpecException, NoSuchProviderException {
 		if (passwordHashType != null && salt != null && len_salt > 0) {
 			byte s[] = new byte[len_salt];
@@ -152,20 +152,20 @@ public class P2PJPAKESecretMessageExchanger extends Agreement {
 	}
 
 	private static byte[] hashMessage(AbstractMessageDigest messageDigest, char password[], byte[] salt,
-			int offset_salt, int len_salt, PasswordHashType passwordHashType, int cost)
+			int offset_salt, int len_salt, PasswordHashType passwordHashType, byte cost)
 			throws NoSuchAlgorithmException, InvalidKeySpecException, NoSuchProviderException {
 		if (salt != null && len_salt > 0) {
 			byte s[] = new byte[len_salt];
 			System.arraycopy(salt, offset_salt, s, 0, len_salt);
 			byte[] res = passwordHashType.hash(password, s, cost, passwordHashType.getDefaultHashLengthBytes());
-			return hashMessage(messageDigest, res, 0, res.length, null, -1, -1, null, 0);
+			return hashMessage(messageDigest, res, 0, res.length, null, -1, -1, null, (byte)0);
 		} else {
 			byte[] res = new byte[password.length * 2];
 			for (int i = 0; i < password.length; i++) {
 				res[i * 2] = (byte) (password[i] & 0xFF);
 				res[i * 2 + 1] = (byte) ((password[i] >> 8) & 0xFF);
 			}
-			return hashMessage(messageDigest, res, 0, res.length, null, -1, -1, null, 0);
+			return hashMessage(messageDigest, res, 0, res.length, null, -1, -1, null, (byte)0);
 		}
 
 	}
